[RT22] Remove FRS and PMR446 Frequency Tables
This driver was coded in a way that erasing a memory row would preselect the
appropriate FRS or PMR446 frequency, bandwidth and transmit power level. This
would similar to how the OEM programming software functions (actually the OEM
software has the frequencies, bandwidth and power level settings fixed so that
they cannot be edited).

While getting ready to submit test images for these models, I rediscovered that
the CHIRP test suite regards this as a memory that can't be erased and reports
an error.

This patch removes the frequency tables and other code related to this so that
the test suite will pass when tested with the related "image" files.

related to #6467