[H777] Rework Serial Timeout
The previous patch has broken uploads for the classic BF-888s radios and
newer Retevis H777 models. The constanstant switching of the timeout to .5
before each block is sent and back to .25 after each block is sent on a
Windows computer creates "noise" that causes the apparently "faster"
radios to mis the "ack" returned by the radio which results in the upload
failing.

This patch resolves the issue by setting the serial timeout to .5 and leaving
it remain that way for all serial activity.

fixes #8197